Yet again, signed comparisons to literals has been fixed.   * compare9.c - an exhaustive '<=' compare to char and int literals   * compare10.c - an exhaustive '>=' compare to char and int literals
git-svn-id: https://sdcc.svn.sourceforge.net/svnroot/sdcc/trunk/sdcc@1710 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ src/regression/compare7.c
@@ -12,8 +12,8 @@ unsigned char dummy = 0;
 unsigned char result = 0;
 
 //bit bit0 = 0;
-long int0 = 0;
-long int1 = 0;
+int int0 = 0;
+int int1 = 0;
 signed char char0 = 0;
 signed char char1 = 0;
 char long0 = 0;
